app/requirements.txt
streamlit==1.44.1
gspread==5.12.4
google-api-python-client==2.169.0
oauth2client==4.1.3
pandas==2.2.3
python-dotenv==1.0.0